requirements.txt
numpy>=1.19.5
tensorflow>=2.6.0
networkx>=2.6.3
matplotlib>=3.4.3
PyYAML>=6.0
scipy>=1.7.1
pandas>=1.3.3
tqdm>=4.62.3